+++ L24-webServer/src/main/java/ru/otus/crm/service/DbServiceClientImpl.java
@@ -61,24 +61,4 @@ public class DbServiceClientImpl implements DBServiceClient {
             return clientList;
        });
     }
-
-    public static DBServiceClient getDBServiceClient(){
-        var configuration = new Configuration().configure(HIBERNATE_CFG_FILE);
-
-        var dbUrl = configuration.getProperty("hibernate.connection.url");
-        var dbUserName = configuration.getProperty("hibernate.connection.username");
-        var dbPassword = configuration.getProperty("hibernate.connection.password");
-
-        new MigrationsExecutorFlyway(dbUrl, dbUserName, dbPassword).executeMigrations();
-
-        var sessionFactory = HibernateUtils.buildSessionFactory(configuration, Client.class, Address.class, Phone.class);
-
-        var transactionManager = new TransactionManagerHibernate(sessionFactory);
-///
-        var clientTemplate = new DataTemplateHibernate<>(Client.class);
-///
-
-        var dbServiceClient = new DbServiceClientImpl(transactionManager, clientTemplate);
-        return dbServiceClient;
-    }
 }
