+++ L24-webServer/src/main/resources/db/migration/V1__initial_migration.sql
@@ -1,0 +1,35 @@
+create sequence hibernate_sequence start with 1 increment by 1;
+create sequence client_sequence start with 1 increment by 1;
+create sequence address_sequence start with 1 increment by 1;
+create sequence phone_sequence start with 1 increment by 1;
+
+create table address
+(
+    id int8 generated by default as identity,
+     street varchar(255),
+      primary key (id)
+);
+create table client
+(
+    id int8 not null,
+     name varchar(255),
+      address_id int8,
+       primary key (id)
+);
+create table phone
+(
+    id int8 not null,
+     number varchar(255),
+      client_id int8,
+       primary key (id)
+);
+
+insert into address (street) values ('Matrosova');
+insert into address (street) values ('Bogoslovskaya');
+insert into client (address_id, name, id) values (1, 'Ilya', 1);
+insert into client (address_id, name, id) values (2, 'Kolya', 2);
+
+insert into phone (client_id, number, id) values (1, 123, 1);
+insert into phone (client_id, number, id) values (1, 1234, 2);
+insert into phone (client_id, number, id) values (2, 234, 3);
+insert into phone (client_id, number, id) values (2, 2345, 4);
