+++ L24-webServer/src/main/resources/hibernate.cfg.xml
@@ -15,7 +15,7 @@
     <property name="hibernate.show_sql">true</property>
     <property name="hibernate.generate_statistics">true</property>
 
-    <property name="hibernate.hbm2ddl.auto">update</property>
+    <property name="hibernate.hbm2ddl.auto">validate</property>
     <property name="hibernate.enable_lazy_load_no_trans">false</property>
 
   </session-factory>
